move all the files
