+++ static/app/views/settings/seer/aiFeaturesAreDisabledBanner.tsx
