+++ static/app/views/settings/seer/noActiveSeerSubscriptionBanner.tsx
