+++ static/app/views/settings/seer/projectDetails/autofixAgent.tsx
