+++ static/app/views/settings/seer/projectDetails/autofixRepositoriesItem.tsx
