+++ static/app/views/settings/seer/projectDetails/autofixRepositoriesList.stories.tsx
@@ -2,15 +2,14 @@ import * as Storybook from 'sentry/stories';
 import {PickProject} from 'sentry/stories/pickProject';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useProjectFromSlug} from 'sentry/utils/useProjectFromSlug';
-
-import {AutofixRepositories} from './autofixRepositoriesList';
+import {AutofixRepositoriesList} from 'sentry/views/settings/seer/projectDetails/autofixRepositoriesList';
 
 export default Storybook.story('AutofixRepositories', story => {
   story('Default', () => {
     function Example({projectSlug}: {projectSlug: string}) {
       const organization = useOrganization();
       const project = useProjectFromSlug({organization, projectSlug});
-      return project && <AutofixRepositories canWrite project={project} />;
+      return project && <AutofixRepositoriesList canWrite project={project} />;
     }
 
     return (
