+++ static/app/views/settings/seer/projectTable/seerProjectTable.tsx
