+++ static/app/views/settings/seer/repoDefaults/repoDefaultsForm.tsx
@@ -10,8 +10,7 @@ import {t, tct} from 'sentry/locale';
 import {DEFAULT_CODE_REVIEW_TRIGGERS} from 'sentry/types/integrations';
 import type {Organization} from 'sentry/types/organization';
 import {fetchMutation} from 'sentry/utils/queryClient';
-
-import {useCanWriteSettings} from 'getsentry/views/seerAutomation/components/useCanWriteSettings';
+import {useCanWriteSettings} from 'sentry/views/settings/seer/useCanWriteSettings';
 
 const schema = z.object({
   autoEnableCodeReview: z.boolean(),
