+++ static/app/views/settings/seer/repoDetails/repoDetailsForm.tsx
