+++ static/app/views/settings/seer/seerAutomationDefault.tsx
