+++ static/app/views/settings/seer/seerAutomationProjectList.tsx
