+++ static/gsApp/views/seerAutomation/advanced.tsx
@@ -13,9 +13,9 @@ import type {Organization} from 'sentry/types/organization';
 import {fetchMutation} from 'sentry/utils/queryClient';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {SettingsPageHeader} from 'sentry/views/settings/components/settingsPageHeader';
+import {useCanWriteSettings} from 'sentry/views/settings/seer/useCanWriteSettings';
 
 import {SeerSettingsPageBanners} from 'getsentry/views/seerAutomation/components/seerSettingsPageBanners';
-import {useCanWriteSettings} from 'getsentry/views/seerAutomation/components/useCanWriteSettings';
 
 const schema = z.object({
   enableSeerCoding: z.boolean(),
